# runner/openapi.reporter.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "AISOC Runner",
    "version": "0.1.0",
    "description": "AISOC Runner executes tool calls against SOCGateway."
  },
  "servers": [
    {
      "url": "https://REPLACE_ME"
    }
  ],
  "components": {
    "securitySchemes": {
      "runnerKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-aisoc-runner-key"
      }
    }
  },
  "security": [
    {
      "runnerKey": []
    }
  ],
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "healthz",
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tools/execute": {
      "post": {
        "operationId": "toolsExecute",
        "summary": "Execute a tool by name",
        "parameters": [
          {
            "in": "header",
            "name": "x-aisoc-agent",
            "required": true,
            "schema": {
              "type": "string",
              "default": "reporter"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tool_name", "arguments"],
                "properties": {
                  "tool_name": {
                    "type": "string",
                    "enum": ["kql_query", "list_incidents", "get_incident", "update_incident", "add_incident_comment"]
                  },
                  "arguments": {
                    "type": "object",
                    "description": "Tool arguments. For kql_query: {query: string, timespan?: string}. workspaceId may be provided by some clients and will be ignored. For get_incident/update_incident/add_incident_comment: prefer {id: <incident GUID or ARM id>} or {incidentNumber: int}."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {"type": "object"}
                  }
                }
              }
            }
          },
          "400": {"description": "Bad request"},
          "401": {"description": "Missing token"},
          "403": {"description": "Forbidden"}
        }
      }
    }
  }
}